UserId attribution added to the functions on TodoItemService.

+++ TodoList.Business/Services/TodoItemService.cs
@@ -26,10 +26,11 @@ namespace TodoList.Business.Services
             _unitOfWork = unitOfWork;
         }
 
-        public async Task<TodoItemDto> CreateAsync(CreateTodoItemDto createTodo)
+        public async Task<TodoItemDto> CreateAsync(CreateTodoItemDto createTodo, Guid userId)
         {
             var todoItem = _mapper.Map<TodoItem>(createTodo);
             todoItem.CreatedDate = DateTime.Now;
+            todoItem.UserId = userId;
             todoItem.IsCompleted = false;
 
             _unitOfWork.TodoItemRepository.Add(todoItem);
@@ -38,24 +39,24 @@ namespace TodoList.Business.Services
             return _mapper.Map<TodoItemDto>(todoItem);
         }
 
-        public async Task DeleteAsync(Guid id)
+        public async Task DeleteAsync(Guid id, Guid userId)
         {
             var todo = await _unitOfWork.TodoItemRepository.GetByIdAsync(id);
 
-            if (todo != null)
+            if (todo != null && todo.UserId == userId)
             {
                 _unitOfWork.TodoItemRepository.Delete(todo);
                 await _unitOfWork.CompleteAsync();
             }
         }
 
-        public async Task<PagedResponse<TodoItemDto>> GetAllAsync(TodoQueryParameters queryParameters)
+        public async Task<PagedResponse<TodoItemDto>> GetAllAsync(TodoQueryParameters queryParameters, Guid userId)
         {
             Expression<Func<TodoItem, bool>> filter = null;
 
             if (queryParameters.Status.HasValue)
             {
-                filter = t => t.IsCompleted == queryParameters.Status.Value;
+                filter = t => t.IsCompleted == queryParameters.Status.Value && t.UserId == userId;
             }
 
             var totalCount = await _unitOfWork.TodoItemRepository.CountAsync(filter);
@@ -69,17 +70,23 @@ namespace TodoList.Business.Services
             return new PagedResponse<TodoItemDto>(todoItemDtos, currentPage: queryParameters.PageNumber, pageSize: queryParameters.PageSize, totalCount: totalCount);
         }
 
-        public async Task<TodoItemDto> GetByIdAsync(Guid id)
+        public async Task<TodoItemDto> GetByIdAsync(Guid id, Guid userId)
         {
             var todoItem = await _unitOfWork.TodoItemRepository.GetByIdAsync(id);
+
+            if (todoItem == null || todoItem.UserId != userId)
+            {
+                todoItem = null;
+            }
+
             return _mapper.Map<TodoItemDto>(todoItem);
         }
 
-        public async Task UpdateAsync(Guid id, UpdateTodoItemDto todoItem)
+        public async Task UpdateAsync(Guid id, UpdateTodoItemDto todoItem, Guid userId)
         {
             var todo = await _unitOfWork.TodoItemRepository.GetByIdAsync(id);
 
-            if (todo != null)
+            if (todo != null && todo.UserId == userId)
             {
                 _mapper.Map(todoItem, todo);
                 todo.UpdatedDate = DateTime.Now;
